+++ src/index.ts
@@ -108,6 +108,8 @@ export * from './field/field.component';
 
 import { CenterComponent } from './center/center.component';
 
+import { LoadingContainer } from './loading-container/loading-container.component';
+
 import {
   MatToolbarModule,
   MatMenuModule,
@@ -116,6 +118,7 @@ import {
   MatCardModule,
   MatIconModule,
   MatRippleModule,
+  MatProgressSpinnerModule,
 } from '@angular/material';
 
 const components = [
@@ -170,6 +173,7 @@ const components = [
   FieldComponent,
   CenterComponent,
   NestableFormDirective,
+  LoadingContainer,
 ];
 
 @NgModule({
@@ -182,6 +186,7 @@ const components = [
     MatCardModule,
     MatIconModule,
     MatRippleModule,
+    MatProgressSpinnerModule,
   ],
   declarations: components,
   exports: [...components],