+++ src/loading-container/loading-container.component.ts
@@ -1,0 +1,41 @@
+import {Component, Input, OnInit} from '@angular/core';
+
+const DEFAULT_LOADING_DELAY_MS = 400;
+
+@Component({
+    selector: 'ign-loading-container',
+    styleUrls: ['./loading-container.scss'],    
+    template: `
+    <div [ngSwitch]="loading">
+	    <div *ngSwitchCase="false">
+	      <ng-content></ng-content>
+      </div>
+      <div *ngSwitchCase="true">
+        <div [ngSwitch]="pastDelay">
+          <div *ngSwitchCase="false">
+          </div>
+          <div *ngSwitchCase="true" class="spinner-container">
+            <mat-spinner></mat-spinner>
+          </div>
+        </div>
+    </div>`
+})
+
+export class LoadingContainer implements OnInit {
+    @Input() loading: boolean;
+    pastDelay: boolean;
+    private delay: number;
+    private delayTimeout: any;
+
+    constructor() {
+      this.delay = DEFAULT_LOADING_DELAY_MS;
+      this.loading = true;
+      this.pastDelay = false;
+    }
+
+    ngOnInit() {
+      this.delayTimeout = setTimeout(() => {
+        this.pastDelay = true;
+      }, this.delay);
+    }
+}